+++ R/helpers.R
@@ -57,3 +57,39 @@
     if (length(novel) > 0) args <- c(args, x[novel])
     args
 }
+
+#' @importFrom S4Vectors mcols 'mcols<-'
+#' @importFrom InteractionSet GInteractions
+#' @keywords internal
+.giFromSj <- function(sj, tx_col, rank_col) {
+
+    stopifnot("site" %in% colnames(mcols(sj)))
+    stopifnot(all(sj$site %in% c("donor", "acceptor")))
+    stopifnot(is.numeric(mcols(sj)[[rank_col]]))
+
+    site <- c() ## R CMD check error avoidance
+    dnr <- subset(sj, site == "donor")
+    dnr$sj <- mcols(dnr)[[rank_col]]
+    dnr_ids <- paste(mcols(dnr)[[tx_col]], dnr$sj)
+
+    acc <- subset(sj, site == "acceptor")
+    acc$sj <- mcols(acc)[[rank_col]] - 1
+    acc_ids <- paste(mcols(acc)[[tx_col]], acc$sj)
+
+    dnr_to_acc <- match(dnr_ids, acc_ids)
+    acc_to_dnr <- match(acc_ids, dnr_ids)
+    if (any(is.na(dnr_to_acc))) stop("NA values when mapping junctions")
+
+    cols <- setdiff(colnames(mcols(dnr)), c("site", rank_col))
+    gi <- GInteractions(
+        anchor1 = granges(dnr)[dnr_to_acc],
+        anchor2 = granges(acc)[acc_to_dnr]
+    )
+    ## Single columns tend to misbehave a little when adding mcols
+    if (length(cols) > 1) {
+        mcols(gi) <- mcols(dnr)[dnr_to_acc, cols]
+    } else {
+        mcols(gi)[[cols]] <- mcols(dnr)[dnr_to_acc, cols]
+    }
+    gi
+}